pyproject.toml
[project]
name = "pkuautovenues-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pycryptodome>=3.23.0",
    "requests>=2.32.5",
    "rsa>=4.9.1",
]
